chore: refactor test to avoid influence each other
