+++ integration-tests/common-test-code/src/test/java/com/vaadin/flow/quarkus/it/ServiceTest.java
@@ -26,10 +26,6 @@ import org.junit.jupiter.api.TestInstance;
 
 import com.vaadin.flow.quarkus.it.service.BootstrapCustomizer;
 import com.vaadin.flow.quarkus.it.service.ServiceBean;
-import com.vaadin.flow.quarkus.it.service.ServiceView;
-import com.vaadin.flow.server.SessionDestroyEvent;
-import com.vaadin.flow.server.SessionInitEvent;
-import com.vaadin.flow.server.UIInitEvent;
 
 @QuarkusTest
 @TestInstance(TestInstance.Lifecycle.PER_CLASS)
@@ -69,33 +65,6 @@ public class ServiceTest extends AbstractCdiTest {
         Assertions.assertEquals(id, getText("service-id"));
         count = getCount(ServiceBean.class.getName());
         Assertions.assertEquals(1, count);
-    }
-
-    @Test
-    public void sessionInitEventObserved() throws IOException {
-        String initCounter = SessionInitEvent.class.getSimpleName();
-        assertCountEquals(0, initCounter);
-        getDriver().manage().deleteAllCookies();
-        open();
-        assertCountEquals(1, initCounter);
-    }
-
-    @Test
-    public void sessionDestroyEventObserved() throws IOException {
-        String destroyCounter = SessionDestroyEvent.class.getSimpleName();
-        assertCountEquals(0, destroyCounter);
-        open();
-        assertCountEquals(0, destroyCounter);
-        click(ServiceView.EXPIRE);
-        assertCountEquals(1, destroyCounter);
-    }
-
-    @Test
-    public void uiInitEventObserved() throws IOException {
-        String uiInitCounter = UIInitEvent.class.getSimpleName();
-        assertCountEquals(0, uiInitCounter);
-        open();
-        assertCountEquals(1, uiInitCounter);
     }
 
 }
